refactor(listings): migrate administration and aggregation services
