+++ core/src/domains/listings/application/ad/AdDuplicateService.ts
@@ -1,10 +1,10 @@
 import mongoose, { ClientSession } from 'mongoose';
 import { createHash } from 'crypto';
-import { getListingRepository } from '../composition/listings';
-import { ListingFilter } from '../domains/listings';
-import DuplicateEvent from '../models/DuplicateEvent';
+import { getListingRepository } from '../../../../composition/listings';
+import { ListingFilter } from '../../../../domains/listings';
+import DuplicateEvent from '../../../../models/DuplicateEvent';
 import { LISTING_STATUS } from '@esparex/contracts';
-import logger from '../utils/logger';
+import logger from '../../../../utils/logger';
 
 // ─────────────────────────────────────────────────
 // TYPES
