+++ core/src/domains/listings/application/ad/AdImageService.ts
@@ -7,11 +7,11 @@
 
 import mongoose, { ClientSession } from 'mongoose';
 import { createHash } from 'crypto';
-import AdImage from '../models/AdImage';
-import { deleteFromS3Url } from '../utils/s3';
-import { processSingleImage } from '../utils/imageProcessor';
-import logger from '../utils/logger';
-import { AppError } from '../utils/AppError';
+import AdImage from '../../../../models/AdImage';
+import { deleteFromS3Url } from '../../../../utils/s3';
+import { processSingleImage } from '../../../../utils/imageProcessor';
+import logger from '../../../../utils/logger';
+import { AppError } from '../../../../utils/AppError';
 
 // ─────────────────────────────────────────────────
 // TYPES & CONSTANTS
@@ -365,7 +365,7 @@ export const findImageDuplicates = async (
 // CANONICAL METHODS (Ex-Shims)
 // ─────────────────────────────────────────────────
 
-import { generatePresignedUploadUrl, PresignedUploadResult } from '../utils/s3';
+import { generatePresignedUploadUrl, PresignedUploadResult } from '../../../../utils/s3';
 
 /**
  * Generates a presigned URL for direct S3 upload from the client.
