+++ core/src/domains/listings/application/ad/ad/AdAggregationService.ts
@@ -1,1 +1,1 @@
-export * from '../adAggregation';
+export * from '../../../../../services/adAggregation';
