+++ core/src/domains/listings/application/ad/ad/_shared/adFilterHelpers.ts
@@ -6,12 +6,12 @@
  */
 
 import mongoose, { PipelineStage } from 'mongoose';
-import BlockedUser from '../../../models/BlockedUser';
+import BlockedUser from '../../../../../../models/BlockedUser';
 import { type ListingTypeValue } from '@esparex/contracts';
-import logger from '../../../utils/logger';
+import logger from '../../../../../../utils/logger';
 
 
-import AdminMetrics from '../../../models/AdminMetrics';
+import AdminMetrics from '../../../../../../models/AdminMetrics';
 
 // ─────────────────────────────────────────────────
 // TYPES & CONSTANTS
