+++ core/src/domains/listings/application/ad/ad/_shared/adServiceBase.ts
@@ -8,40 +8,40 @@
 
 export { default as mongoose } from 'mongoose';
 export type { PipelineStage } from 'mongoose';
-export { default as Ad } from '../../../models/Ad';
-export { default as Category } from '../../../models/Category';
-export { default as Brand } from '../../../models/Brand';
-export { default as Model } from '../../../models/Model';
-export { default as Business } from '../../../models/Business';
-export { default as Report } from '../../../models/Report';
-export { default as BlockedUser } from '../../../models/BlockedUser';
-export { default as SparePart } from '../../../models/SparePart';
-export { default as ServiceType } from '../../../models/ServiceType';
-export { default as CatalogRequest } from '../../../models/CatalogRequest';
-export { serializeDoc } from '../../../utils/serialize';
-export { normalizeLocationResponse } from '../../location/LocationNormalizer';
-export { touchLocationSearchAnalytics } from '../../location/LocationAnalyticsService';
-export { buildGeoNearStage, normalizeGeoInput } from '../../../utils/mongoGeoUtils';
-export { normalizeAdStatus } from '../../lifecycle/AdStatusService';
-export { buildAdFilterFromCriteria } from '../../../utils/adFilterHelper';
-export type { AdFilterCriteria } from '../../../utils/adFilterHelper';
-export { getCache, setCache, getMultiCache, setMultiCache, CACHE_KEYS } from '../../../utils/redisCache';
-export { buildPublicAdFilter } from '../../../utils/FeedVisibilityGuard';
+export { default as Ad } from '../../../../../../models/Ad';
+export { default as Category } from '../../../../../../models/Category';
+export { default as Brand } from '../../../../../../models/Brand';
+export { default as Model } from '../../../../../../models/Model';
+export { default as Business } from '../../../../../../models/Business';
+export { default as Report } from '../../../../../../models/Report';
+export { default as BlockedUser } from '../../../../../../models/BlockedUser';
+export { default as SparePart } from '../../../../../../models/SparePart';
+export { default as ServiceType } from '../../../../../../models/ServiceType';
+export { default as CatalogRequest } from '../../../../../../models/CatalogRequest';
+export { serializeDoc } from '../../../../../../utils/serialize';
+export { normalizeLocationResponse } from '../../../../../../services/location/LocationNormalizer';
+export { touchLocationSearchAnalytics } from '../../../../../../services/location/LocationAnalyticsService';
+export { buildGeoNearStage, normalizeGeoInput } from '../../../../../../utils/mongoGeoUtils';
+export { normalizeAdStatus } from '../../../../../../services/lifecycle/AdStatusService';
+export { buildAdFilterFromCriteria } from '../../../../../../utils/adFilterHelper';
+export type { AdFilterCriteria } from '../../../../../../utils/adFilterHelper';
+export { getCache, setCache, getMultiCache, setMultiCache, CACHE_KEYS } from '../../../../../../utils/redisCache';
+export { buildPublicAdFilter } from '../../../../../../utils/FeedVisibilityGuard';
 export { LISTING_TYPE, type ListingTypeValue } from '@esparex/contracts';
-export { default as logger } from '../../../utils/logger';
-export { default as RankingTelemetry } from '../../../models/RankingTelemetry';
+export { default as logger } from '../../../../../../utils/logger';
+export { default as RankingTelemetry } from '../../../../../../models/RankingTelemetry';
 export { v4 as uuidv4 } from 'uuid';
-export { escapeRegExp } from '../../../utils/stringUtils';
+export { escapeRegExp } from '../../../../../../utils/stringUtils';
 export {
     buildAdSortStage as buildAdSortStageFromHelper,
     extractLocationIdFromAd,
     normalizeAdImagesForResponse,
-} from '../../adQuery/AdQueryHelpers';
-export type { SortStage } from '../../adQuery/AdQueryHelpers';
+} from '../../../queries/adQuery/AdQueryHelpers';
+export type { SortStage } from '../../../queries/adQuery/AdQueryHelpers';
 export { LISTING_STATUS } from '@esparex/contracts';
-export { FeatureFlag, isEnabled } from '../../../config/featureFlags';
-export { default as AdminMetrics } from '../../../models/AdminMetrics';
-export { isBusinessPublishedStatus } from '../../../utils/businessStatus';
+export { FeatureFlag, isEnabled } from '../../../../../../config/featureFlags';
+export { default as AdminMetrics } from '../../../../../../models/AdminMetrics';
+export { isBusinessPublishedStatus } from '../../../../../../utils/businessStatus';
 export {
     getBlockedSellerIds,
     recordListingTypeCompatMetric,
