+++ core/src/domains/listings/application/aggregation/adAggregation/types.ts
