+++ core/src/services/AdCreationService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdCreationService";
-export { default } from "../domains/listings/application/ad/AdCreationService";
