+++ core/src/services/AdDuplicateService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdDuplicateService";
-export { default } from "../domains/listings/application/ad/AdDuplicateService";
