+++ core/src/services/AdEngagementService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdEngagementService";
-export { default } from "../domains/listings/application/ad/AdEngagementService";
