+++ core/src/services/AdImageService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdImageService";
-export { default } from "../domains/listings/application/ad/AdImageService";
