+++ core/src/services/AdMutationService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/mutations/AdMutationService";
-export { default } from "../domains/listings/application/mutations/AdMutationService";
