+++ core/src/services/AdOrchestrator.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdOrchestrator";
-export { default } from "../domains/listings/application/ad/AdOrchestrator";
