+++ core/src/services/AdValidationService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/AdValidationService";
-export { default } from "../domains/listings/application/ad/AdValidationService";
