+++ core/src/services/ListingMutationService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/mutations/ListingMutationService";
-export { default } from "../domains/listings/application/mutations/ListingMutationService";
