+++ core/src/services/ListingSubmissionPolicy.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/domain/policies/ListingSubmissionPolicy";
-export { default } from "../domains/listings/domain/policies/ListingSubmissionPolicy";
