+++ core/src/services/SavedAdService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/saved/SavedAdService";
-export { default } from "../domains/listings/application/saved/SavedAdService";
