+++ core/src/services/SparePartListingService.ts
@@ -1,2 +1,1 @@
 export * from "../domains/listings/application/ad/SparePartListingService";
-export { default } from "../domains/listings/application/ad/SparePartListingService";
