+++ core/src/services/adminListings/index.ts
@@ -1,5 +1,1 @@
-export * from './types';
-export * from './helpers';
-export * from './queries';
-export * from './mutations';
-export * from './bulk';
+export * from "../../domains/listings/application/moderation/adminListings";
